Rename: 컨트롤러-뷰 사이 DTO를 둠에 따라 패키지를 이동한다.
