+++ src/main/java/org/prgms/springbootbasic/controller/voucher/VoucherController.java
@@ -3,7 +3,7 @@ package org.prgms.springbootbasic.controller.voucher;
 import lombok.extern.slf4j.Slf4j;
 import org.prgms.springbootbasic.domain.VoucherType;
 import org.prgms.springbootbasic.domain.voucher.Voucher;
-import org.prgms.springbootbasic.domain.voucher.dto.VoucherRequestDto;
+import org.prgms.springbootbasic.controller.voucher.dto.VoucherRequestDto;
 import org.prgms.springbootbasic.exception.EntityNotFoundException;
 import org.prgms.springbootbasic.service.VoucherService;
 import org.springframework.stereotype.Controller;
