+++ src/main/java/org/prgms/springbootbasic/controller/voucher/dto/VoucherRequestDto.java
@@ -1,4 +1,4 @@
-package org.prgms.springbootbasic.domain.voucher.dto;
+package org.prgms.springbootbasic.controller.voucher.dto;
 
 public record VoucherRequestDto (String voucherId, String voucherPolicy, long discountDegree) {
 }
